Save a deviceDescriptor instead of macaddress when adding a device
This is a bit of a hack, because the variable "macaddress" throughout
the app no longer stores a simple mac address. It's a string that
includes the macaddress and a device type identifier.

Also TBD, the name presented to the user in this commit is the above string.
If we decide it would be helpful to the user to show her the device type as
part of the default name, some reformatting needs to be done to show the longer
name properly. Or, we need to split the string and just display the mac address.